[TASK] Cleanup options define and loading
Centralize all the options to the options module, and then leverage
import_opt or import_group to include specific options within the
current scope.

All logic for setting up options, logging, etc. centralized within
the service module reduce number of individual imports and calls
from a few to a single function call.
