+++ etc/seedbox/gen.conf
@@ -1,8 +1,4 @@
 [DEFAULT]
 output_file = etc/seedbox/seedbox.conf.sample
 namespace = seedbox
-namespace = seedbox.db
-namespace = seedbox.process
-namespace = seedbox.task
-namespace = seedbox.torrent
