+++ seedbox/cli.py
@@ -4,46 +4,23 @@
 The main program that is the entry point for the SeedboxManager application.
 Provides the ability to configure and start up processing.
 """
-import logging
 import os
+import tempfile
 
-from lockfile import pidlockfile
-from oslo_config import cfg
+import lockfile
 
-from seedbox import db
-from seedbox import logext as logmgr
-from seedbox import options
 from seedbox import process
+from seedbox import service
 
-LOG = logging.getLogger(__name__)
 
-
+@lockfile.locked(os.path.join(tempfile.gettempdir(), __package__), timeout=10)
 def main():
     """Entry point for seedmgr"""
 
     # processes all command-line inputs that control how we execute
     # logging, run mode, etc.; and we get a handle back to access
     # the info
-    options.initialize()
+    service.prepare_service()
 
-    # configure our logging
-    logmgr.configure()
-    cfg.CONF.log_opt_values(LOG, logging.DEBUG)
-
-    try:
-        # need to create a lock to make sure multiple instances do not start
-        # at the same time because we are running as a cron.
-        with pidlockfile.PIDLockFile(os.path.join(cfg.CONF.config_dir,
-                                     'seedmgr.lock'), timeout=10):
-
-            # setup database and initialize connection
-            db.dbapi(cfg.CONF)
-
-            # time to start processing
-            process.start()
-
-    except pidlockfile.LockTimeout as lockerr:
-        # if we have managed timeout, it means there is another instance
-        # already running so we will simply bow out and let the existing
-        # one still run.
-        LOG.info('Already running; if not delete lock file: %s', lockerr)
+    # time to start processing
+    process.start()
