+++ seedbox/db/__init__.py
@@ -10,19 +10,7 @@ from seedbox.db import api
 LOG = logging.getLogger(__name__)
 DB_ENGINE_NAMESPACE = 'seedbox.db'
 
-OPTS = [
-    cfg.StrOpt('connection',
-               default='sqlite:///$config_dir/torrent.db',
-               help='The connection string used to connect to the database'),
-    cfg.IntOpt('idle_timeout',
-               default=3600,
-               help='Timeout before idle sql connections are reaped'),
-    cfg.IntOpt('connection_debug',
-               default=0,
-               help='Verbosity of SQL debugging information. 0=None, 100=All'),
-]
-
-cfg.CONF.register_opts(OPTS, 'database')
+cfg.CONF.import_group('database', 'seedbox.options')
 
 _DBAPI = None
 
@@ -49,23 +37,3 @@ def dbapi(conf=cfg.CONF):
         _DBAPI = api.DBApi(_get_connection(conf))
         _DBAPI.shrink_db()
     return _DBAPI
-
-
-def list_opts():
-    """Returns a list of oslo_config options available in the library.
-
-    The returned list includes all oslo_config options which may be registered
-    at runtime by the library.
-
-    Each element of the list is a tuple. The first element is the name of the
-    group under which the list of elements in the second element will be
-    registered. A group name of None corresponds to the [DEFAULT] group in
-    config files.
-
-    The purpose of this is to allow tools like the Oslo sample config file
-    generator to discover the options exposed to users by this library.
-
-    :returns: a list of (group_name, opts) tuples
-    """
-    from seedbox.common import tools
-    return tools.make_opt_list([OPTS], 'database')
