+++ seedbox/tasks/filesync.py
@@ -12,42 +12,8 @@ from seedbox.tasks import subprocessext
 
 LOG = logging.getLogger(__name__)
 
-OPTS = [
-    cfg.BoolOpt('dryrun',
-                default=False,
-                help='rsync dryrun option'),
-    cfg.BoolOpt('verbose',
-                default=False,
-                help='rsync verbose option'),
-    cfg.BoolOpt('progress',
-                default=False,
-                help='rsync progress option'),
-    cfg.BoolOpt('perms',
-                default=True,
-                help='rsync perms option'),
-    cfg.BoolOpt('delayupdates',
-                default=True,
-                help='rsync delayupdates option'),
-    cfg.BoolOpt('recursive',
-                default=True,
-                help='rsync recursive option'),
-    cfg.StrOpt('chmod',
-               default='ugo+rwx',
-               help='rsync chmod option'),
-    cfg.StrOpt('identity',
-               help='rsync-ssh identity option (ssh key)'),
-    cfg.StrOpt('port',
-               default='22',
-               help='rsync-ssh port'),
-    cfg.StrOpt('remote_user',
-               help='User name on remote system (ssh)'),
-    cfg.StrOpt('remote_host',
-               help='Host name/IP Address of remote system'),
-    cfg.StrOpt('remote_path',
-               help='rsync destination path'),
-]
-
-cfg.CONF.register_opts(OPTS, group='tasks_filesync')
+cfg.CONF.import_group('tasks', 'seedbox.options')
+cfg.CONF.import_group('tasks_filesync', 'seedbox.options')
 
 
 class SyncFile(base.BaseTask):
