+++ setup.cfg
@@ -54,10 +54,6 @@ seedbox.db =
 
 oslo.config.opts =
     seedbox = seedbox.options:list_opts
-    seedbox.db = seedbox.db:list_opts
-    seedbox.process = seedbox.process:list_opts
-    seedbox.task = seedbox.tasks:list_opts
-    seedbox.torrent = seedbox.torrent:list_opts
 
 [build_sphinx]
 builders=html
@@ -102,8 +98,8 @@ autodoc_exclude_modules =
         seedbox.tests.test
         seedbox.tests.test_cli
         seedbox.tests.test_gen_config
-        seedbox.tests.test_logext
         seedbox.tests.test_options
+        seedbox.tests.test_service
         seedbox.tests.test_version
         seedbox.tests.torrent.test_bencode
         seedbox.tests.torrent.test_loader
